\name{aid}
\alias{aid}
\title{ monthly payment for federal program }
\description{ monthly payment for federal program }
\usage{data(aid)}
\format{
  The format is:
 Named num [1:51]  57.2 253.5 114.2  68.2 199.6 ...
 - attr(*, "names")= chr [1:51] "Alabama" "Alaska" "Arizona" "Arkansas" ...
}
\source{
  From Kitchen's Exploring Statistics
}
\examples{
data(aid)
hist(aid)
}
\keyword{datasets}
