\name{answers}
\alias{answers}
\title{
Answers to selected problems
}
\description{
Opens pdf file containing answers to selected problems
}
\usage{
answers()
}
\value{Called for its side-effect of opening a pdf
}
\examples{
## answers()
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ utilities }
