\name{bumpers}
\alias{bumpers}
\title{ Bumper repair costs for various automobiles}
\description{ bumper repair costs }
\usage{data(bumpers)}
\format{
  Price in dollars to repair a bumper.
}
\source{
  From \emph{Exploring Statistics}, Duxbury Press, 1998, L. Kitchens.
}
\examples{
data(bumpers)
stem(bumpers)
}
\keyword{datasets}
