\name{cancer}
\alias{cancer}
\title{ cancer survival times  }
\description{ cancer survival times }
\usage{data(cancer)}
\format{
  The format is:
  The format is:
List of 5 numeric components stomach, bronchus, colon, ovary and breast
}
\source{
  Taken  from L. Kitchens, \emph{Exploring Statistics}, Duxbury Press, 1997.
}
\examples{
data(cancer)
boxplot(cancer)
}
\keyword{datasets}
