\name{central.park.cloud}
\alias{central.park.cloud}
\docType{data}
\title{ Type of day in Central Park, NY May 2003}
\description{
  The type of day in May 2003 in Central Park, NY
}
\usage{data(central.park.cloud)}
\format{
  A factor with levels \code{clear},\code{partly.cloudy} and \code{cloudy}.
}
\source{
  This type of data, and much more, is available from
  \url{http://www.noaa.gov}.
}
\examples{
data(central.park.cloud)
table(central.park.cloud)
}
\keyword{datasets}
