\name{ceo2013}
\alias{ceo2013}
\docType{data}
\title{
CEO compensation in 2013
}
\description{
 Data on top 200 CEO compensations in the year 2013
}
\usage{data(ceo2013)}
\format{
 a data frame.
}
\source{
Scraped from \url{http://www.nytimes.com/interactive/2013/06/30/business/executive-compensation-tables.html?ref=business}
}
\examples{
data(ceo2013)
}
\keyword{datasets}
