\name{demos}
\alias{demos}
\title{
Provide menu for possible shiny demonstrations
}
\description{
Provides a menu to open one of the provided demonstrations which use \pkg{shiny} for animation.
}
\usage{
demos()
}
\details{
User must have installed \pkg{shiny} prior to usage. As \pkg{shiny} has some dependencies that don't always work, this package is not a dependency of \pkg{UsingR}.
}
\value{No return value, when called a web page opens. Use Ctrl-C (or equivalent) in terminal to return to an interactive session.
}
\examples{
## demos()
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ dynamic }
