\name{exec.pay}
\alias{exec.pay}
\title{ Direct compensation for 199 United States CEOs in the year 2000  }
\description{ Direct compensation for 199 United States CEOs in the year 2000
in units of \$10,000.  }
\usage{data(exec.pay)}
\format{
  A numeric vector with 199 entries each measuring compensation in 10,000s
  of dollars.
}
\source{
\emph{New York Times} Business section 04/01/2001. See also \url{http://www.aflcio.org}.
}
\examples{
data(exec.pay)
hist(exec.pay)
}
\keyword{datasets}
