\name{firstchi}
\alias{firstchi}
\title{Age of mother at birth of first child }
\description{Age of mother at birth of first child }
\usage{data(firstchi)}
\format{
  The format is:
 num [1:87] 30 18 35 22 23 22 36 24 23 28 ...
}
\source{
  From \emph{Exploring Statistics},  L. Kitchens, Duxbury Press, 1998.
}
\examples{
data(firstchi)
hist(firstchi)
}
\keyword{datasets}
