\name{gap}
\alias{gap}
\docType{data}
\title{ Sales data for the Gap}
\description{
  Sales data for the Gap from Jan
}
\usage{data(gap)}
\format{
  The format is a ts object storing data from June 2002 through June 2005.
}
\source{
  http://home.businesswire.com
}
\examples{
data(gap)
monthplot(gap)
}
\keyword{datasets}
