\name{goalspergame}
\alias{goalspergame}
\docType{data}
\title{ Goals per game in NHL}
\description{
  Goals per game in NHL
}
\usage{data(goalspergame)}
\format{
  The format is:
 mts [1:53, 1:4] 6 6 6 6 6 6 6 6 6 6 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:4] "n.teams" "n.games" "n.goals" "gpg"
 - attr(*, "tsp")= num [1:3] 1946 1998    1
 - attr(*, "class")= chr [1:2] "mts" "ts"
}
\source{
  Off internet site. Forgot which.
}
\examples{
data(goalspergame)
}
\keyword{datasets}
