\name{homework}
\alias{homework}
\title{Homework averages for Private and Public schools }
\usage{data(homework)}
\description{Homework averages for Private and Public schools 
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{Private}{a numeric vector}
    \item{Public}{a numeric vector}
  }
}
\source{
  This is from Kitchens Exploring Statistics
}
\examples{
data(homework)
boxplot(homework)
}
\keyword{datasets}
