\name{HUMMER}
\alias{HUMMER}
\docType{data}
\title{Deliveries of new HUMMER vehicles}
\description{
  Gives monthly delivery numbers for new HUMMER vehicles from June 2003 through
  February 2006. During July, August, and September 2005 there was an
  Employee Pricing Incentive.
}
\usage{data(HUMMER)}
\format{
  The format is:
 Time-Series [1:33] from 2003 to 2006: 2493 2654 2987 2837 3157 2837 3157 1927 2141 2334 ...
}
\source{
  Compiled from delivery data avalailble at
  http://www.gm.com/company/investor_information/sales_prod/hist_sales.html
}
\examples{
data(HUMMER)
plot(HUMMER)
}
\keyword{datasets}
