\name{income_percentiles}
\alias{income_percentiles}
\docType{data}
\title{
Top percentiles of U.S. income
}
\description{
Top percentiles of U.S. income
}
\usage{data(income_percentiles)}
\format{
	A data frame with \code{Year} and various percentile (90th, 95th, ...)
}
\source{Not available}
\examples{
data(income_percentiles)
}
\keyword{datasets}
