\name{malpract}
\alias{malpract}
\title{ malpractice settlements  }
\description{ malpractice settlements  }
\usage{data(malpract)}
\format{
  The format is:
 num [1:17] 760 380 125 250 2800 450 100 150 2000 180 ...
}
\source{
  From Kitchens Exploring Statistics
}
\examples{
data(malpract)
boxplot(malpract)
}
\keyword{datasets}
