\name{Medicare}
\alias{Medicare}
\docType{data}
\title{
Sample from "Medicare Provider Charge Data"
}
\description{
Sample from "Medicare Provider Charge Data"
}
\usage{data(Medicare)}
\format{
  A data frame with 10000 observations and data for on billings for procedures at many different hospitals.
}
\source{
http://www.cms.gov/Research-Statistics-Data-and-Systems/Statistics-Trends-and-Reports/Medicare-Provider-Charge-Data/index.html
}
\references{
This data came from
http://www.cms.gov/Research-Statistics-Data-and-Systems/Statistics-Trends-and-Reports/Medicare-Provider-Charge-Data/index
and was referenced in the article
\url{http://www.nytimes.com/2013/05/08/business/hospital-billing-varies-wildly-us-data-shows.html},
as retrieved on 5/8/2013.
}
\examples{
data(Medicare)
}
\keyword{datasets}
