\name{movie_data_2011}
\alias{movie_data_2011}
\docType{data}
\title{Movie data for 2011 by weekend}
\description{Movie data for 2011 by weekend}
\usage{data(movie_data_2011)}
\format{
A data frame with variables \code{Previous} (previous weekend rank), \code{Movie} (title), \code{Distributor}, \code{Genre}, \code{Gross} (per current weekend), \code{Change} (change from previous week), \code{Theaters} (number of theaters), \code{TotalGross} (total gross to date), \code{Days} (days out), \code{weekend} (weekend of report)
}
\source{Scraped from pages such as \url{http://www.the-numbers.com/box-office-chart/weekend/2011/04/29}
}
\examples{
data(movie_data_2011)
}
\keyword{datasets}
