\name{nisdc}
\alias{nisdc}
\docType{data}
\title{
NISCD
}
\description{
A data frame measuring daily sea-ice extent from 1978 until 2013.
}
\usage{data(nisdc)}
\format{
  A data frame measuring daily sea-ice extent from 1978 until 2013
}
\source{
ftp://sidads.colorado.edu/DATASETS/NOAA/G02135/north/daily/data/NH_seaice_extent_final.csv
and
ftp://sidads.colorado.edu/DATASETS/NOAA/G02135/north/daily/data/NH_seaice_extent_nrt.csv (now offline).
}
\references{
See the blog post \url{https://www.r-bloggers.com/2012/08/arctic-sea-ice-at-lowest-levels-since-observations-began/} for a description and nice script to play with.
}
\keyword{datasets}
