\name{ObamaApproval}
\alias{ObamaApproval}
\docType{data}
\title{
 Approval ratings for President Obama
}
\description{
A collection of approval ratings for President Obama spanning a duration from early 2010 to the summer of 2013.
}
\usage{data(ObamaApproval)}
\format{
  A data frame 7 variables.
}
\source{
 Scraped on 7-5-13 from \url{http://www.realclearpolitics.com/epolls/other/president_obama_job_approval-1044.html}
}
\examples{
data(ObamaApproval)
}
\keyword{datasets}
