\name{pi2000}
\alias{pi2000}
\title{first 2000 digits of pi }
\description{first 2000 digits of pi }
\usage{data(pi2000)}
\format{
  The format is:
 num [1:2000] 3 1 4 1 5 9 2 6 5 3 ...
}
\source{
  Generated by \emph{Mathematica}, \url{http://www.wolfram.com}.
}
\examples{
data(pi2000)
chisq.test(table(pi2000))
}
\keyword{datasets}
