\name{primes}
\alias{primes}
\docType{data}
\title{ Primes numbers less than 2003}
\description{
  Prime numbers between 1 and 2003.
}
\usage{data(primes)}
\format{
  The format is:
 num [1:304] 2 3 5 7 11 13 17 19 23 29 ...
}
\source{
  Generated using \url{http://www.rsok.com/~jrm/printprimes.html}.
}
\examples{
data(primes)
diff(primes)
}
\keyword{datasets}
