\name{puerto}
\alias{puerto}
\title{Incomes for Puerto Rican immigrants to Miami }
\description{Incomes for Puerto Rican immigrants to Miami }
\usage{data(puerto)}
\format{
  The format is:
 num [1:50] 150 280 175 190 305 380 290 300 170 315 ...
}
\source{
  From Kitchens Exploring Statistics
}
\examples{
data(puerto)
hist(puerto)
}
\keyword{datasets}
