\name{rat}
\alias{rat}
\title{Survival times of 20 rats exposed to radiation}
\description{Survival times of 20 rats exposed to radiation}
\usage{data(rat)}
\format{
  The format is:
 num [1:20] 152 152 115 109 137 88 94 77 160 165 ...
}
\source{
  From Kitchents Exploring Statistics
}
\examples{
data(rat)
hist(rat)
}
\keyword{datasets}
