\name{slc}
\alias{slc}
\title{ Sodium-Lithium countertransport }
\description{ Sodium-Lithium countertransport }
\usage{data(slc)}
\format{
  The format is:
 num [1:190] 0.467 0.430 0.192 0.192 0.293 ...
}
\source{
  From Kitchens' Exploring Statistics
}
\examples{
data(slc)
hist(slc)
}
\keyword{datasets}
