\name{snacks}
\alias{snacks}
\docType{data}
\title{
Snack data from the USDA
}
\description{ subset of SR26 data on nutrients compiled by the USDA.
}
\usage{data(snacks)}
\format{A data frame with some nutrition variables
}

\source{
This data came from the SR26 data set found at \url{http://www.ars.usda.gov/Services/docs.htm?docid=8964}.
}
\examples{
data(snacks)
}
\keyword{datasets}
