\name{south}
\alias{south}
\title{ Murder rates for 30 Southern US cities}
\description{ Murder rates for 30 Southern US cities}
\usage{data(south)}
\format{
  The format is:
 num [1:30] 12 10 10 13 12 12 14 7 16 18 ...
}
\source{
From Kitchens' Exploring Statistics
}
\examples{
data(south)
hist(south)
}
\keyword{datasets}
