\name{wchomes}
\alias{wchomes}
\docType{data}
\title{
 A random sample of Wake County, North Carolina residential real estate plots
}
\description{
This data set comes from a JSE article
 \url{http://www.amstat.org/publications/jse/v16n3/woodard.pdf} by
 Roger Woodard. The data is described by: The information for this
 data set was taken from a Wake County, North Carolina real estate
 database. Wake County is home to the capital of North Carolina,
 Raleigh, and to Cary. These cities are the fifteenth and eighth
 fastest growing cities in the USA respectively, helping Wake County
 become the ninth fastest growing county in the country. Wake County
 boasts a 31.18% growth in population since 2000, with a population
 of approximately 823,345 residents. This data includes 100 randomly
 selected residential properties in the Wake County registry denoted
 by their real estate ID number. For each selected property, 11
 variables are recorded. These variables include year built, square
 feet, adjusted land value, address, et al.
}
\usage{data(wchomes)}
\format{a data frame}
\source{
 \url{http://www.amstat.org/publications/jse/v16n3/woodard.xls}
}
\references{ \url{http://www.amstat.org/publications/jse/v16n3/woodard.pdf}
}
\examples{
data(wchomes)
}
\keyword{datasets}
