\name{wellbeing}
\alias{wellbeing}
\docType{data}
\title{
What makes us happy?
}
\description{
Correlated data on what makes us happy
}
\usage{data(wellbeing)}
\format{A data frame with data about what makes people happy (well being) along with several other covariates
}
\source{Found from \url{http://prcweb.co.uk/lab/what-makes-us-happy/}.
}
\references{\url{http://prcweb.co.uk/lab/what-makes-us-happy/} and \url{http://www.nationalaccountsofwellbeing.org/}
}
\examples{
data(wellbeing)
}
\keyword{datasets}
