% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getAnswer.R
\name{getAnswer}
\alias{getAnswer}
\title{function to get answer to problem}
\usage{
getAnswer(chapter = NULL, problem = NULL)
}
\arguments{
\item{chapter}{which chapter}

\item{problem}{which problem}
}
\value{
opens web page to answer
}
\description{
Returns answers for the first edition.
}

